requirements/test.txt
-r ./common.txt

gunicorn==20.1.0
factory_boy==3.2.1
psycopg2==2.9.2 --no-binary psycopg2
pytest==6.2.5
pytest-django==4.5.2